+++ CHANGELOG.md
@@ -9,6 +9,8 @@ CHANGES:
 * Added a new script ssm/backup-ssm-to-s3.py. ([@TGWolf][])
 * Added a new script ec2/backup-all-ebs-volumes.sh. ([@TGWolf][])
 * Added a new script ec2/list-availability-zones.py. ([@TGWolf][])
+* Added a new script ec2/list-all-ec2-instance-types.py. ([@TGWolf][])
+
 ## 1.0.6 (November 22, 2018)
 
 CHANGES: